refactor(web): promote theme controls
Move the shared toggle component and hook out of navigation and expose them through a dedicated theme entry point.
