+++ apps/web/src/components/Theme/index.ts
@@ -1,0 +1,2 @@
+export * from './ThemeToggle';
+export * from './useThemeToggle';
